pyproject.toml
[project]
name = "rio-jobs"
version = "0.2.0"
description = "A job scheduler for the Rio pure Python web framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["rio-ui>=0.11.0, <0.12.0"]
authors = [{ name = "Jakob Pinterits", email = "jakob.pinterits@gmail.com" }]
license.file = "LICENSE.txt"
keywords = ["jobs", "tasks", "scheduler", "web framework", "rio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[dependency-groups]
dev = ["pre-commit>=3.7.1"]

[project.urls]
homepage = "https://github.com/rio-labs/rio-jobs"
repository = "https://github.com/rio-labs/rio-jobs"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
